整理结构体

+++ app/models/t_admin_role.go
@@ -5,7 +5,7 @@ import "github.com/revel/revel"
 
 //AdminRole model
 type AdminRole struct {
-	Id       int    `xorm:"not null pk autoincr INT(11)"`
+	Id       int64  `xorm:"not null pk autoincr INT(11)"`
 	RoleDes  string `xorm:"VARCHAR(20)"`
 	RoleType int    `xorm:"INT(11)"`
 }
