+++ app/models/t_comment.go
@@ -6,11 +6,11 @@ import "errors"
 
 //Comment model.
 type Comment struct {
-	Id         int       `xorm:"not null pk autoincr INT(11)"`
+	Id         int64     `xorm:"not null pk autoincr INT(11)"`
 	Content    string    `xorm:"TEXT"`
-	Auther     int       `xorm:"INT(11)"`
-	BlogId     int       `xorm:"INT(11)"`
-	CommentId  int       `xorm:"INT(11)"`
+	Auther     int64     `xorm:"INT(11)"`
+	BlogId     int64     `xorm:"INT(11)"`
+	CommentId  int64     `xorm:"INT(11)"`
 	CreateTime time.Time `xorm:"DATETIME"`
 	Status     int       `xorm:"INT(1)"`
 }
